refactor: moving the DA layer to DA folder
