+++ .vscode/settings.json
@@ -5,6 +5,7 @@
         "codegen",
         "czvf",
         "darwin",
+        "dataavailability",
         "dbname",
         "deepmap",
         "devnet",
@@ -52,6 +53,7 @@
         "sqlx",
         "sslmode",
         "stretchr",
+        "syscoin",
         "unmarshal",
         "versioninfo",
         "wangyoucao",
