+++ internal/dataavailability/http_custom_error.go
@@ -1,4 +1,4 @@
-package model
+package dataavailability
 
 import "strconv"
